# backend/src/main/java/com/hospedaya/backend/controller/ComentarioController.java
package com.hospedaya.backend.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import io.swagger.v3.oas.annotations.parameters.RequestBody;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/comentarios")
@Tag(name = "Comentarios", description = "Gestión de comentarios")
public class ComentarioController {

    @Operation(summary = "Listar comentarios")
    @ApiResponses({
            @ApiResponse(responseCode = "200", description = "Comentarios obtenidos"),
            @ApiResponse(responseCode = "400", description = "Petición inválida"),
            @ApiResponse(responseCode = "404", description = "No se encontraron comentarios")
    })
    @GetMapping
    public ResponseEntity<?> listarComentarios() {
        return ResponseEntity.ok().build();
    }

    @Operation(summary = "Crear comentario")
    @RequestBody(
            description = "Datos del comentario",
            required = true,
            content = @Content(mediaType = "application/json",
                    examples = @ExampleObject(
                            value = "{ \"usuarioId\": 1, \"alojamientoId\": 3, \"texto\": \"Muy buen lugar\", \"calificacion\": 5 }"
                    )
            )
    )
    @ApiResponses({
            @ApiResponse(responseCode = "201", description = "Comentario creado"),
            @ApiResponse(responseCode = "400", description = "Datos inválidos"),
            @ApiResponse(responseCode = "404", description = "Usuario o alojamiento no encontrado")
    })
    @PostMapping
    public ResponseEntity<?> crearComentario() {
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }
}