# backend/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Sistema de gestión de alojamientos
  description: API para gestionar alojamientos, reservas, usuarios y servicios relacionados.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /alojamientos:
    get:
      summary: Obtener todos los alojamientos
      tags:
        - Alojamiento
      responses:
        "200":
          description: Lista de alojamientos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlojamientoDTO"
              example:
                - id: 1
                  nombre: "Casa Rural Los Pinos"
                  descripcion: "Una casa rural acogedora en el campo"
                  direccion: "Calle Falsa 123, Pueblo, País"
                  precioPorNoche: 75.50
                  anfitrionId: 1
                - id: 2
                  nombre: "Apartamento en la ciudad"
                  descripcion: "Apartamento moderno cerca del centro"
                  direccion: "Av. Principal 456, Ciudad"
                  precioPorNoche: 120.00
                  anfitrionId: 2
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron alojamientos

    post:
      summary: Crear un nuevo alojamiento
      tags:
        - Alojamiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlojamientoDTO"
            example:
              nombre: "Cabaña en el bosque"
              descripcion: "Cabaña rústica rodeada de naturaleza"
              direccion: "Km 5, vía al bosque, Pueblo"
              precioPorNoche: 90.00
              anfitrionId: 3
      responses:
        "201":
          description: Alojamiento creado con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlojamientoDTO"
              example:
                id: 10
                nombre: "Cabaña en el bosque"
                descripcion: "Cabaña rústica rodeada de naturaleza"
                direccion: "Km 5, vía al bosque, Pueblo"
                precioPorNoche: 90.00
                anfitrionId: 3
        "400":
          description: Datos inválidos para crear alojamiento
        "404":
          description: No se pudo crear el alojamiento
...
  /pagos:
    get:
      summary: Obtener todos los pagos
      tags:
        - Pago
      responses:
        "200":
          description: Lista de pagos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PagoDTO"
              example:
                - id: 1
                  reservaId: 1
                  monto: 500000
                  metodoPago: "Tarjeta de crédito"
                  fechaPago: "2023-07-01"
                - id: 2
                  reservaId: 2
                  monto: 350000
                  metodoPago: "Efectivo"
                  fechaPago: "2023-07-05"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron pagos

    post:
      summary: Registrar un nuevo pago
      tags:
        - Pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagoDTO"
            example:
              reservaId: 1
              monto: 500000
              metodoPago: "Tarjeta de crédito"
              fechaPago: "2023-09-01"
      responses:
        "201":
          description: Pago registrado con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagoDTO"
              example:
                id: 5
                reservaId: 1
                monto: 500000
                metodoPago: "Tarjeta de crédito"
                fechaPago: "2023-09-01"
        "400":
          description: Datos inválidos para registrar el pago
        "404":
          description: Reserva asociada no encontrada
...
  /comentarios:
    get:
      summary: Obtener todos los comentarios
      tags:
        - Comentario
      responses:
        "200":
          description: Lista de comentarios obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComentarioDTO"
              example:
                - id: 1
                  usuarioId: 3
                  alojamientoId: 2
                  reservaId: 5
                  comentario: "Muy buen lugar, cómodo y limpio."
                  calificacion: 5
                  fecha: "2023-07-10"
                - id: 2
                  usuarioId: 4
                  alojamientoId: 1
                  reservaId: 6
                  comentario: "El anfitrión fue muy atento."
                  calificacion: 4
                  fecha: "2023-07-15"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron comentarios

    post:
      summary: Crear un nuevo comentario
      tags:
        - Comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComentarioDTO"
            example:
              usuarioId: 3
              alojamientoId: 2
              reservaId: 5
              comentario: "Excelente ubicación y trato del anfitrión."
              calificacion: 5
              fecha: "2023-09-01"
      responses:
        "201":
          description: Comentario creado con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComentarioDTO"
              example:
                id: 10
                usuarioId: 3
                alojamientoId: 2
                reservaId: 5
                comentario: "Excelente ubicación y trato del anfitrión."
                calificacion: 5
                fecha: "2023-09-01"
        "400":
          description: Datos inválidos para crear comentario
        "404":
          description: Reserva o usuario asociado no encontrado
...
  /favoritos:
    get:
      summary: Obtener todos los favoritos
      tags:
        - Favorito
      responses:
        "200":
          description: Lista de favoritos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoritoDTO"
              example:
                - id: 1
                  usuarioId: 3
                  alojamientoId: 2
                  fechaAgregado: "2023-08-15"
                - id: 2
                  usuarioId: 3
                  alojamientoId: 4
                  fechaAgregado: "2023-08-20"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron favoritos

    post:
      summary: Agregar un alojamiento a favoritos
      tags:
        - Favorito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoritoDTO"
            example:
              usuarioId: 3
              alojamientoId: 5
              fechaAgregado: "2023-09-10"
      responses:
        "201":
          description: Favorito creado con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritoDTO"
              example:
                id: 10
                usuarioId: 3
                alojamientoId: 5
                fechaAgregado: "2023-09-10"
        "400":
          description: Datos inválidos para crear favorito
        "404":
          description: Usuario o alojamiento no encontrado
...
  /imagenes-alojamiento:
    get:
      summary: Obtener todas las imágenes de alojamientos
      tags:
        - ImagenAlojamiento
      responses:
        "200":
          description: Lista de imágenes obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImagenAlojamientoDTO"
              example:
                - id: 1
                  alojamientoId: 2
                  url: "https://miapp.com/images/alojamiento2_1.jpg"
                  descripcion: "Fachada principal del alojamiento"
                - id: 2
                  alojamientoId: 2
                  url: "https://miapp.com/images/alojamiento2_2.jpg"
                  descripcion: "Habitación principal con vista"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron imágenes

    post:
      summary: Subir una nueva imagen para un alojamiento
      tags:
        - ImagenAlojamiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagenAlojamientoDTO"
            example:
              alojamientoId: 3
              url: "https://miapp.com/images/alojamiento3_1.jpg"
              descripcion: "Piscina privada con vista panorámica"
      responses:
        "201":
          description: Imagen agregada con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagenAlojamientoDTO"
              example:
                id: 15
                alojamientoId: 3
                url: "https://miapp.com/images/alojamiento3_1.jpg"
                descripcion: "Piscina privada con vista panorámica"
        "400":
          description: Datos inválidos para crear imagen
        "404":
          description: Alojamiento no encontrado
...
components:
  schemas:
...
    AlojamientoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 2
        nombre:
          type: string
          example: Casa rural
        descripcion:
          type: string
          example: Acogedora casa en el campo con chimenea
        direccion:
          type: string
          example: Calle 123, Salento
        precioPorNoche:
          type: number
          format: double
          example: 75.5
        anfitrionId:
          type: integer
          example: 1
...
    PagoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 20
        reservaId:
          type: integer
          example: 5
        monto:
          type: number
          format: double
          example: 250.75
        metodoPago:
          type: string
          example: TARJETA_CREDITO
        fechaPago:
          type: string
          format: date
          example: 2025-09-10
    ComentarioDTO:
      type: object
      properties:
        id:
          type: integer
          example: 12
        usuarioId:
          type: integer
          example: 1
        alojamientoId:
          type: integer
          example: 2
        texto:
          type: string
          example: Muy buen alojamiento, cómodo y limpio
        calificacion:
          type: integer
          example: 5
    FavoritoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 8
        usuarioId:
          type: integer
          example: 1
        alojamientoId:
          type: integer
          example: 2
    ImagenAlojamientoDTO:
      type: object
      properties:
        id:
          type: integer
          example: 15
        alojamientoId:
          type: integer
          example: 2
        url:
          type: string
          example: http://imagenes.com/casa1.jpg
...